+++ src/lightly_train/_data/task_batch_collation.py
@@ -11,6 +11,7 @@ from typing import Literal
 
 import numpy as np
 import torch
+from albumentations import Compose
 
 from lightly_train._transforms.object_detection_transform import (
     ObjectDetectionTransformArgs,
@@ -81,7 +82,7 @@ class ObjectDetectionCollateFunction(BaseCollateFunction):
     ):
         super().__init__(split, transform_args)
         assert isinstance(transform_args, ObjectDetectionTransformArgs)
-        self.scale_jitter: ScaleJitter | None
+        self.scale_jitter: Compose | None
         if transform_args.scale_jitter is not None:
             if (
                 transform_args.scale_jitter.min_scale is None
@@ -93,15 +94,22 @@ class ObjectDetectionCollateFunction(BaseCollateFunction):
                     transform_args.scale_jitter.min_scale,
                     transform_args.scale_jitter.max_scale,
                 )
-            self.scale_jitter = ScaleJitter(
-                sizes=transform_args.scale_jitter.sizes,
-                target_size=transform_args.image_size,
-                scale_range=scale_range,
-                num_scales=transform_args.scale_jitter.num_scales,
-                divisible_by=transform_args.scale_jitter.divisible_by,
-                p=transform_args.scale_jitter.prob,
-                step_seeding=transform_args.scale_jitter.step_seeding,
-                seed_offset=transform_args.scale_jitter.seed_offset,
+            self.scale_jitter = Compose(
+                [
+                    ScaleJitter(
+                        sizes=transform_args.scale_jitter.sizes,
+                        target_size=transform_args.image_size
+                        if transform_args.scale_jitter.sizes is None
+                        else None,
+                        scale_range=scale_range,
+                        num_scales=transform_args.scale_jitter.num_scales,
+                        divisible_by=transform_args.scale_jitter.divisible_by,
+                        p=transform_args.scale_jitter.prob,
+                        step_seeding=transform_args.scale_jitter.step_seeding,
+                        seed_offset=transform_args.scale_jitter.seed_offset,
+                    )
+                ],
+                bbox_params=transform_args.bbox_params,
             )
         else:
             self.scale_jitter = None
@@ -115,6 +123,7 @@ class ObjectDetectionCollateFunction(BaseCollateFunction):
                     "image": item["image"].permute(1, 2, 0).numpy(),
                     "bboxes": item["bboxes"].numpy(),
                     "classes": item["classes"].numpy(),
+                    "original_size": item["original_size"],
                 }
                 for item in batch
             ]
@@ -147,7 +156,7 @@ class ObjectDetectionCollateFunction(BaseCollateFunction):
 
             # Turn back into torch tensors.
             images = [
-                torch.from_numpy(img).to(torch.float32).permute(2, 0, 1)
+                torch.from_numpy(img).permute(2, 0, 1).to(torch.float32)
                 for img in images
             ]
             bboxes = [torch.from_numpy(bbox).to(torch.float32) for bbox in bboxes]
@@ -158,6 +167,7 @@ class ObjectDetectionCollateFunction(BaseCollateFunction):
                 "image": torch.stack(images),
                 "bboxes": bboxes,
                 "classes": classes,
+                "original_size": [item["original_size"] for item in batch],
             }
             return out_
         else:
@@ -166,5 +176,6 @@ class ObjectDetectionCollateFunction(BaseCollateFunction):
                 "image": torch.stack([item["image"] for item in batch]),
                 "bboxes": [item["bboxes"] for item in batch],
                 "classes": [item["classes"] for item in batch],
+                "original_size": [item["original_size"] for item in batch],
             }
             return out_