+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Object detection metrics tracking and required dependencies.
 - Add support for DINOv3 ConvNeXt models.
 - Automatically download DINOv3 weights.
 - Predict masks for a full dataset with a semantic segmentation model checkpoint with